backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
debugpy==1.8.6
Django==4.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-timezone-field==7.0
djangorestframework==3.15.2
docker==7.1.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
kombu==5.4.2
outcome==1.3.0.post0
packaging==23.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.0.0
pycparser==2.21
PySocks==1.7.1
pytest==8.3.3
pytest-celery==1.1.3
pytest-django==4.9.0
pytest-docker-tools==3.1.3
pytest-mock==3.14.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.31.0
selenium==4.18.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.5.1
tenacity==9.0.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
wsproto==1.2.0
